new: add --validate-response to get-historic-readings
this will validate the response and will err if empty fields are 100% empty or not present in the response. In any other case, it will print the number of rows that have nulls.

it will add a bit of overhead to the task. To deactivate it, remove the --validate-response from the command in the dag.